Move non-conch performers out of conch module, per code review
